Fix soroban fee tests that that check events and refactor the math to make logic clearer

+++ src/transactions/test/SorobanTxTestUtils.h
@@ -60,7 +60,7 @@ ContractEvent makeMintOrBurnEvent(bool isMint, const stellar::Hash& contractId,
                                   int64 amount);
 
 bool validateFeeEvent(ContractEvent const& feeEvent, PublicKey const& feeSource,
-                      int feeCharged);
+                      int64_t feeCharged);
 
 // Creates a valid transaction for uploading provided Wasm.
 // Fills in the valid footprint automatically in case if `uploadResources`
